\name{as.r.default}
\alias{as.r.default}
\title{turn a vector into r code}
\description{This function converts an R vector (ignoring any attributes) into R source code that will produce that vector. The current version can only print the R code out on the console, and is intended for copying and pasting.}
\usage{
\method{as.r}{default}(x, \dots)
}
\arguments{
\item{x}{}
\item{\dots}{}
}

