\name{function}
\alias{mutate.function}
\title{change the body of a function}
\description{The mutate.function method changes the body of a function. The str argument is a character vector, representing the lines of the new function.}
\usage{
\method{mutate}{function}(f, str, \dots)
}
\arguments{
\item{f}{}
\item{str}{}
\item{\dots}{}
}

