\name{mobject}
\alias{mclass}
\alias{mdelete}
\alias{is.MObject}
\alias{print.MObject}
\alias{~}
\alias{~.default}
\alias{~.MObject}
\title{m object system}
\description{deprecated, do not use}
\usage{
mclass (super, \dots, hybrid=TRUE, abstract=FALSE)
mdelete (cid)
is.MObject (obj)
\method{print}{MObject}(obj, \dots)
obj~invocation
}
\arguments{
\item{super}{.}
\item{abstract}{.}
\item{hybrid}{.}
\item{cid}{.}
\item{obj}{.}
\item{invocation}{.}
\item{\dots}{.}
}



