\name{newgenerics}
\alias{aslatex}
\alias{equals}
\alias{clone}
\alias{mutate}
\alias{fit}
\alias{evaluate}
\alias{gf}
\alias{map}
\alias{reset}
\title{New Generics}
\description{New generic functions.}
\usage{
aslatex (\dots)
equals (\dots)
clone (\dots)
mutate (\dots)
fit (\dots)
evaluate (\dots)
gf (\dots)
map (\dots)
reset (\dots)
}
\arguments{
\item{\dots}{}
}

