\name{oospdefaults}
\alias{oospprint.default}
\alias{oospformat.default}
\alias{oospsummary.default}
\alias{oospplot.default}
\alias{oosplines.default}
\alias{oosppredict.default}
\alias{oospfitted.default}
\alias{oospresiduals.default}
\alias{oospweights.default}
\alias{oospaslist.default}
\alias{oospasdataframe.default}
\title{Copycat Defaults}
\description{If no other method is available for the copycat generic, use the standard generic.}
\usage{
\method{oospprint}{default}(\dots)
\method{oospformat}{default}(\dots)
\method{oospsummary}{default}(\dots)
\method{oospplot}{default}(\dots)
\method{oosplines}{default}(\dots)
\method{oosppredict}{default}(\dots)
\method{oospfitted}{default}(\dots)
\method{oospresiduals}{default}(\dots)
\method{oospweights}{default}(\dots)
\method{oospaslist}{default}(\dots)
\method{oospasdataframe}{default}(\dots)
}
\arguments{
\item{\dots}{}
}

