\name{oospgenerics}
\alias{oospprint}
\alias{oospformat}
\alias{oospsummary}
\alias{oospplot}
\alias{oosplines}
\alias{oosppredict}
\alias{oospfitted}
\alias{oospresiduals}
\alias{oospweights}
\alias{oospaslist}
\alias{oospasdataframe}
\title{Copycat Generics}
\description{Copycat generics that do not constrain argument names.}
\usage{
oospprint (\dots)
oospformat (\dots)
oospsummary (\dots)
oospplot (\dots)
oosplines (\dots)
oosppredict (\dots)
oospfitted (\dots)
oospresiduals (\dots)
oospweights (\dots)
oospaslist (\dots)
oospasdataframe (\dots)
}
\arguments{
\item{\dots}{}
}

