\name{pointer}
\alias{pointer}
\alias{map.pointer}
\alias{oospprint.pointer}
\alias{ptrenv}
\alias{length.pointer}
\alias{[.pointer}
\alias{[<-.pointer}
\alias{[[.pointer}
\alias{[[<-.pointer}
\title{Pointer Objects}
\description{Create and Manipulate Pointer Objects}
\usage{
pointer(obj, hash=FALSE)
\method{map}{pointer}(ptr, obj, \dots)
\method{oospprint}{pointer}(ptr, \dots)
ptrenv(ptr)
\method{length}{pointer}(x)
\method{[}{pointer}(ptr, i, \dots)
\method{[}{pointer}(ptr, i, \dots) <- value
\method{[[}{pointer}(ptr, i)
\method{[[}{pointer}(ptr, i, \dots) <- value
}
\arguments{
\item{ptr}{}
\item{x}{}
\item{obj}{}
\item{hash}{}
\item{i}{}
\item{value}{}
\item{\dots}{}
}

