\name{print.ENVIRONMENT}
\alias{print.ENVIRONMENT}
\alias{print.compenv}
\title{print an ENVIRONMENT object}
\description{Print a ENVIRONMENT object like a list. We can also use this function on regular environments if we use the full function name.}
\usage{
\method{print}{ENVIRONMENT}(e, \dots)
\method{print}{compenv}(e, \dots)
}
\arguments{
\item{e}{.}
\item{\dots}{.}
}

