\name{compenv}
\alias{compenv}
\alias{as.compenv}
\alias{is.compenv}
\title{Component Environments}
\description{Create, coerce and test compenv objects.}
\usage{
compenv (\dots, hash=FALSE)
as.compenv (obj)
is.compenv (e)
}
\arguments{
\item{\dots}{Seed objects (each must be named).}
\item{hash}{Whether or not the compenv environment is hashed. Refer to the help for new.env for more information.}
\item{obj}{}
\item{e}{}
}

