\name{method}
\alias{freemethod}
\alias{hypermethod}
\alias{print.freemethod}
\title{method objects}
\description{deprecated, do not use}
\usage{
freemethod (f, \dots)
hypermethod (obj, mid)
\method{print}{freemethod}(f, \dots)
}
\arguments{
\item{f}{.}
\item{obj}{.}
\item{mid}{.}
\item{\dots}{.}
}

