\name{FUNCTION}
\alias{FUNCTION}
\alias{print.FUNCTION}
\alias{$.FUNCTION}
\alias{$<-.FUNCTION}
\title{Enhanced Functions}
\description{Enhanced function objects, refer to the relevant package vignette for more information.}
\usage{
FUNCTION (f, \dots, self.reference=".")
\method{print}{FUNCTION}(f, \dots)
\method{$}{FUNCTION}(x, name)
\method{$}{FUNCTION}(x, name) <- value
}
\arguments{
\item{f}{A function.}
\item{x}{A function.}
\item{self.reference}{The self reference operator.}
\item{name}{.}
\item{value}{.}
\item{\dots}{.}
}



