\name{newgenerics}
\alias{map}
\alias{fit}
\alias{evaluate}
\alias{reset}
\alias{clone}
\alias{mutate}
\alias{gf}
\alias{as.r}
\alias{as.latex}
\title{new generics}
\description{New standard generic functions.}
\usage{
map (\dots)
fit (\dots)
evaluate (\dots)
reset (\dots)
clone (\dots)
mutate (\dots)
gf (\dots)
as.r (\dots)
as.latex (\dots)
}
\arguments{
\item{\dots}{The arguments.}
}

